Refactor Mediator project to use RequestPipeline approach as it more efficient performance-wise.
This approach makes it at least 2 times faster performance-wise, since we rely a bit less in reflection.
Now we only rely on it once - for resolving the request pipeline - whereas previously we relied on it for resolving all the handlers and behaviors for a given request. This means that previously, the more behaviors we had, the slower it could potentially become.